+++ Constants/PathUrlRedirect.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace PlacetoPay\Payments\Constants;
+
+abstract class PathUrlRedirect
+{
+    public const FAILURE = 'placetopay/onepage/failure';
+    public const SUCCESSFUL = 'placetopay/onepage/success';
+    public const PENDING = 'placetopay/onepage/pending';
+}